amazing stuff :clap:

+++ src/core/primitives.hpp
@@ -7,8 +7,8 @@ namespace bolt
 {
     struct vector_2
     {
-        double x = 0.0f;
-        double y = 0.0f;
+        float x = 0.0f;
+        float y = 0.0f;
 
         //vector addition
         [[nodiscard]] inline vector_2 operator + (vector_2 const vec) const
@@ -43,13 +43,13 @@ namespace bolt
         }
 
         // vector doproduct
-        [[nodiscard]] inline double operator | (vector_2 const &vec) const
+        [[nodiscard]] inline float operator | (vector_2 const &vec) const
         {
             return this->x*vec.x + this->y*vec.y;
         }
 
         // multiply by vector
-        [[nodiscard]] inline vector_2 operator * (double const &scalar) const
+        [[nodiscard]] inline vector_2 operator * (float const &scalar) const
         {
             return {
                 this->x * scalar,
@@ -57,13 +57,13 @@ namespace bolt
             };
         }
 
-        inline void operator *= (double const &scalar)
+        inline void operator *= (float const &scalar)
         {
             this->x *= scalar;
             this->y *= scalar;
         }
 
-        [[nodiscard]] inline vector_2 operator / (double const &scalar) const
+        [[nodiscard]] inline vector_2 operator / (float const &scalar) const
         {
             return {
                 this->x / scalar,
@@ -71,20 +71,20 @@ namespace bolt
             };
         }
 
-        inline void operator /= (double const &scalar)
+        inline void operator /= (float const &scalar)
         {
             this->x /= scalar;
             this->y /= scalar;
         }
 
-        [[nodiscard]] inline double length() const
+        [[nodiscard]] inline float length() const
         {
-            return (double)sqrtl((*this) | (*this));
+            return (float)sqrtl((*this) | (*this));
         }
 
         [[nodiscard]] inline vector_2 normalize() const
         {
-            double l = this->length();
+            float l = this->length();
             bool lf = l != 0, lt = l == 0;
             return {
                 ((this->x / l) * lf + this->x * lt),
@@ -100,9 +100,9 @@ namespace bolt
 
     struct vector_3
     {
-        double x = 0.0f;
-        double y = 0.0f;
-        double z = 0.0f;
+        float x = 0.0f;
+        float y = 0.0f;
+        float z = 0.0f;
 
         // vector addition
         [[nodiscard]] inline vector_3 operator + (vector_3 const &vec) const
@@ -141,7 +141,7 @@ namespace bolt
         }
 
         // vector multiply by scalar
-        [[nodiscard]] inline vector_3 operator * (double const &scalar) const
+        [[nodiscard]] inline vector_3 operator * (float const &scalar) const
         {
             return {
                 this->x * scalar,
@@ -151,7 +151,7 @@ namespace bolt
         }
 
         // vector multiply by scalar
-        inline void operator *= (double const &scalar)
+        inline void operator *= (float const &scalar)
         {
             this->x *= scalar;
             this->y *= scalar;
@@ -159,7 +159,7 @@ namespace bolt
         }
 
         // vector dot product
-        [[nodiscard]] inline double operator | (vector_3 const &vec) const
+        [[nodiscard]] inline float operator | (vector_3 const &vec) const
         {
             return this->x*vec.x + this->y*vec.y + this->z*vec.z;
         }
@@ -183,15 +183,15 @@ namespace bolt
         }
 
         // calculate vector length
-        [[nodiscard]] inline double length() const
+        [[nodiscard]] inline float length() const
         {
-            return (double)sqrtl((*this) | (*this));
+            return (float)sqrtl((*this) | (*this));
         }
 
         //normalized vector
         [[nodiscard]] inline vector_3 normalize() const
         {
-            double l = this->length();
+            float l = this->length();
             bool lf = l != 0, lt = l == 0;
             return {
                 ((this->x / l) * lf + this->x * lt),
@@ -208,7 +208,7 @@ namespace bolt
     
     struct matrix_3
     {
-        double m[3][3] {
+        float m[3][3] {
             1.0f, 0.0f, 0.0f,
             0.0f, 1.0f, 0.0f,
             0.0f, 0.0f, 1.0f
@@ -228,7 +228,7 @@ namespace bolt
 
     struct matrix_4
     {
-        double m[4][4] {
+        float m[4][4] {
             1.0f, 0.0f, 0.0f, 0.0f,
             0.0f, 1.0f, 0.0f, 0.0f,
             0.0f, 0.0f, 1.0f, 0.0f,
@@ -246,7 +246,7 @@ namespace bolt
             };
         }
 
-        [[nodiscard]] static matrix_4 rotation_x(double deg)
+        [[nodiscard]] static matrix_4 rotation_x(float deg)
         {
             return {
                 1.0f, 0.0f, 0.0f, 0.0f,
@@ -256,7 +256,7 @@ namespace bolt
             };
         }
 
-        [[nodiscard]] static matrix_4 rotation_y(double deg)
+        [[nodiscard]] static matrix_4 rotation_y(float deg)
         {
             return {
                 cos(deg), 0.0f, sin(deg), 0.0f,
@@ -266,7 +266,7 @@ namespace bolt
             };
         }
 
-        [[nodiscard]] static matrix_4 rotation_z(double deg)
+        [[nodiscard]] static matrix_4 rotation_z(float deg)
         {
             return {
 	        	cos(deg), -sin(deg), 0.0f, 0.0f, 
@@ -347,21 +347,21 @@ namespace bolt
         [[nodiscard]] inline vector_3 operator* (vector_3 vec) const
         {
             // Now set some simple names for the vec
-            double x = vec.x;
-            double y = vec.y;
-            double z = vec.z;
+            float x = vec.x;
+            float y = vec.y;
+            float z = vec.z;
 
             // Multiply the vec against each part of the 1st column, then add together
-            double resX = (x * this->m[0][0]) + (y * this->m[1][0]) + (z * this->m[2][0]) + (this->m[3][0]);
+            float resX = (x * this->m[0][0]) + (y * this->m[1][0]) + (z * this->m[2][0]) + (this->m[3][0]);
 
             // Multiply the vec against each part of the 2nd column, then add together
-            double resY = (x * this->m[0][1]) + (y * this->m[1][1]) + (z * this->m[2][1]) + (this->m[3][1]);
+            float resY = (x * this->m[0][1]) + (y * this->m[1][1]) + (z * this->m[2][1]) + (this->m[3][1]);
 
             // Multiply the vec against each part of the 3rd column, then add together
-            double resZ = (x * this->m[0][2]) + (y * this->m[1][2]) + (z * this->m[2][2]) + (this->m[3][2]);
+            float resZ = (x * this->m[0][2]) + (y * this->m[1][2]) + (z * this->m[2][2]) + (this->m[3][2]);
 
             // Multiply the vec against each part of the 4th column, then add together
-            double resW = (x * this->m[0][3]) + (y * this->m[1][3]) + (z * this->m[2][3]) + (this->m[3][3]);
+            float resW = (x * this->m[0][3]) + (y * this->m[1][3]) + (z * this->m[2][3]) + (this->m[3][3]);
 
             if (resW != 0) {
                 resX /= resW; resY /= resW; resZ /= resW;
